+++ packages.nix
@@ -2,36 +2,39 @@
   lib,
   pkgs,
   system,
-  rockchipOsConfigAarch64,
-  rockchipInstallerConfigAarch64,
-  osConfigAarch64,
-  installerConfigAarch64,
-  installerConfigX86,
+  myLib,
 }:
 
 let
   image-gnome =
-    (rockchipOsConfigAarch64 system ./devices/pinebook-pro.nix ./gnome.nix).config.system.build.sdImage;
+    (myLib.rockchipOsConfigAarch64 system ./hosts/nixos/PineBookPro ./hosts/common/optional/gnome.nix)
+    .config.system.build.sdImage;
   image-plasma =
-    (rockchipOsConfigAarch64 system ./devices/pinebook-pro.nix ./plasma.nix)
+    (myLib.rockchipOsConfigAarch64 system ./hosts/nixos/PineBookPro ./hosts/common/optional/plasma.nix)
     .config.system.build.sdImage;
   image-pinetab2-gnome =
-    (rockchipOsConfigAarch64 system ./devices/pinetab2.nix ./gnome.nix).config.system.build.sdImage;
+    (myLib.rockchipOsConfigAarch64 system ./hosts/nixos/PineTab2 ./hosts/common/optional/gnome.nix)
+    .config.system.build.sdImage;
   image-pinetab2-plasma =
-    (rockchipOsConfigAarch64 system ./devices/pinetab2.nix ./plasma.nix).config.system.build.sdImage;
-  uboot = (rockchipOsConfigAarch64 system ./devices/pinebook-pro.nix { }).config.rockchip.uBoot;
-  uboot-pinetab2 = (rockchipOsConfigAarch64 system ./devices/pinetab2.nix { }).config.rockchip.uBoot;
+    (myLib.rockchipOsConfigAarch64 system ./hosts/nixos/PineTab2 ./hosts/common/optional/plasma.nix)
+    .config.system.build.sdImage;
+  uboot = (myLib.rockchipOsConfigAarch64 system ./hosts/nixos/PineBookPro { }).config.rockchip.uBoot;
+  uboot-pinetab2 =
+    (myLib.rockchipOsConfigAarch64 system ./hosts/nixos/PineTab2 { }).config.rockchip.uBoot;
   image-installer-pinebookpro =
-    (rockchipInstallerConfigAarch64 system ./devices/pinebook-pro.nix).config.system.build.sdImage;
+    (myLib.rockchipInstallerConfigAarch64 system ./hosts/nixos/PineBookPro).config.system.build.sdImage;
   image-installer-pinetab2 =
-    (rockchipInstallerConfigAarch64 system ./devices/pinetab2.nix).config.system.build.sdImage;
+    (myLib.rockchipInstallerConfigAarch64 system ./hosts/nixos/PineTab2).config.system.build.sdImage;
   image-generic-aarch64 =
-    (osConfigAarch64 system ./devices/generic-aarch64.nix ./gnome.nix).config.system.build.sdImage;
+    (myLib.osConfigAarch64 system ./hosts/nixos/GenericAarch64 ./hosts/common/optional/gnome.nix)
+    .config.system.build.sdImage;
   image-installer-generic-aarch64 =
-    (installerConfigAarch64 system ./devices/generic-aarch64.nix).config.system.build.sdImage;
-  image-installer-x86pc = (installerConfigX86 { }).config.system.build.isoImage;
-  image-installer-x86pc-gnome = (installerConfigX86 ./gnome.nix).config.system.build.isoImage;
-  image-installer-x86pc-plasma = (installerConfigX86 ./plasma.nix).config.system.build.isoImage;
+    (myLib.installerConfigAarch64 system ./hosts/nixos/GenericAarch64).config.system.build.sdImage;
+  image-installer-x86pc = (myLib.installerConfigX86 { }).config.system.build.isoImage;
+  image-installer-x86pc-gnome =
+    (myLib.installerConfigX86 ./hosts/common/optional/gnome.nix).config.system.build.isoImage;
+  image-installer-x86pc-plasma =
+    (myLib.installerConfigX86 ./hosts/common/optional/plasma.nix).config.system.build.isoImage;
 in
 {
   inherit